requirements.txt
django-registration-redux==2.13
django-tree-queries==0.19.0
sentry-sdk==2.12.0
pygments==2.18.0
django>=5.0.7
dj-database-url==2.2.0
django-storage-url==0.12.0
whitenoise==5.2.0
#boto3==2.49.0
uwsgi==2.0.26
#django-material==1.7.4
django-extensions==3.2.3
#django-webmention==2.0.2
is-safe-url==1.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-ordered-model==3.7.4
django-extra-views==0.14.0
psycopg==3.2.1
django-multiselectfield==0.1.13
django-countries==7.6.1
#plotly==5.18.0
#kaleido==0.2.1
#logdna==1.18.12
django-taggit==6.0.0
Jinja2==3.1.4
logica==1.3.1415926535
pyparsing==3.1.2
